Implementacion de filtros

+++ database/migrations/2021_11_17_215224_create_desgloses_table.php
@@ -16,7 +16,7 @@ class CreateDesglosesTable extends Migration
         Schema::create('desgloses', function (Blueprint $table) {
             $table->foreignId('facturas_id');
             $table->foreignId('producto_id');
-            $table->float('cantidad');
+            $table->integer('cantidad');
             $table->float('pre_tot');
             $table->timestamps();
         });
